Add an MGL unit to the dataset (#4)
This is an MGL console with one of the lower serial numbers. The chip numbers are also slightly different (lower numbers) than the other MGL unit.

To get the site generator to run I needed to add the Mitsumi transformer chip to the matcher.

I found the LCD's row driver but am not sure where the column driver is. I didn't want to displace the screen.

The screen's backing plate has the number 8213 stamped on it but I didn't list this as the screen's label since it is missing from the component database. Additionally, the LCD's ribbon cable has a numeric stamp on it (T80303) and a part number (M1085) but I didn't add those fields since they are not part of the schema.

Tested by building and running the site locally.